Move timestamp module to root crate
