+++ modules/src/application/ics20_fungible_token_transfer/relay_application_logic/send_transfer.rs
@@ -5,7 +5,7 @@ use crate::handler::HandlerOutput;
 use crate::ics04_channel::handler::send_packet::send_packet;
 use crate::ics04_channel::packet::Packet;
 use crate::ics04_channel::packet::PacketResult;
-use crate::ics24_host::timestamp::Timestamp;
+use crate::timestamp::Timestamp;
 
 pub(crate) fn send_transfer<Ctx>(
     ctx: &Ctx,
