+++ modules/src/ics04_channel/error.rs
@@ -6,7 +6,7 @@ pub type Error = anomaly::Error<Kind>;
 use super::packet::Sequence;
 use crate::ics04_channel::channel::State;
 use crate::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
-use crate::ics24_host::timestamp::Timestamp;
+use crate::timestamp::Timestamp;
 use crate::{ics02_client, Height};
 
 #[derive(Clone, Debug, Error, Eq, PartialEq)]
