+++ modules/src/ics24_host/mod.rs
@@ -5,5 +5,4 @@ pub use path::{ClientUpgradePath, Path, IBC_QUERY_PATH, SDK_UPGRADE_QUERY_PATH};
 pub mod error;
 pub mod identifier;
 mod path;
-pub mod timestamp;
 pub mod validate;
